feat: add detect-schema-drift workflow to detect drift of query.types.json
